Package renamed
Renamed the package to my github username.
